# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/packages/{registry}/{package}/owners": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all owners of a specific package",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Package Ownership"
                ],
                "summary": "Get package owners",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Registry type (e.g., npm, nuget, maven)",
                        "name": "registry",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Package name",
                        "name": "package",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package owners retrieved successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "owners": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Package not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve package owners",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}